requirements.txt
# Sayer - Requirements File
# Developed by Saudi Linux (SaudiLinux7@gmail.com)

# Core dependencies
requests>=2.28.1
python-nmap>=0.7.1
colorama>=0.4.6
argparse>=1.4.0
tqdm>=4.64.1
pyyaml>=6.0
pretty-tables>=0.7.2
bs4>=0.0.1
beautifulsoup4>=4.11.1
lxml>=4.9.1
whois>=0.9.22
dnspython>=2.2.1
fake-useragent>=1.1.1
python-whois>=0.8.0

# Web scanning dependencies
requests-html>=0.10.0
requests-futures>=1.0.0
requests-ntlm>=1.1.0
requests-toolbelt>=0.10.1

# Network scanning dependencies
scapy>=2.4.5
netaddr>=0.8.0

# Reporting dependencies
jinja2>=3.1.2
matplotlib>=3.6.2
pandas>=1.5.1
openpyxl>=3.0.10

# API integrations
shodan>=1.28.0
censys>=2.1.8

# GUI dependencies (optional)
pyqt5>=5.15.7